Delay loading of offer descriptions
Gatsby preloads the data needed for a page when you hover a link to
that page. In the case of an offer map, the "data needed" was quite
large in size - 825 KB gzipped in the case of Quebec apartments.

The preloading behavior is very nice, however. I played with quite a few
options to reduce the size of the payload, and settled on fetching offer
descriptions with AJAX instead of server-side rendering them. Additionally, we
now whitelist offer attributes; reducing the size of the API response even
further.

Concretely, the 825 KB initial request for the Quebec apartments has became a
122 KB initial request + a 441 KB AJAX request afterwards for the offer
descriptions.

Followup work (number that follows is possible reduction in payload size):
    * Skip offers for which we don't have an exact address (28%)
    * Skip offers that are outside X kilometers of the city center (6%)
    * Round latitude/longitude to meter accuracy (0.8%)
    * Do not use an array of objects in the response; instead, push the
      properties directly into an array. Would require some frontend changes.
      (9%)