Add filtering proof of concept
Performance is completely unacceptable. Typing "1" into the min price triggers
a re-render that takes ~3 seconds and blocks the main thread. Looking in the
performance tab, the majority of the work seems to be in the "setIcon" method
of Leaflet. Given that filtering works perfectly fine with stand-alone Leaflet,
I have to assume that the slowness stems from react-leaflet's wrapping of
Leaflet.