Fix regression on language-based redirects
Technically, "/" is not a page on this site. All pages are instead
prefixed by "fr" or "en" and I've set up redirects from "/" to the
appropriate language (this applies to all pages, not just the home
page).

Unfortunately, there seems to have been a regression in
gatsby-plugin-netlify. As brought up in
https://github.com/gatsbyjs/gatsby/issues/18220, if multiple redirects
are created from the same page, all but the latest will be ignored, even
if the redirects have different options.  Previously, I was creating the
following redirects for each page:

    /quebec/  /en/quebec/  302
    /quebec/  /fr/quebec/  302 Language=fr

This let me redirect people to French/English based on their language
header. However, because of the linked issue, this is no longer
possible - only the French redirect is being generated. This is
effectively breaking redirects for all non-French speakers.

As a temporary fix, I am going to give up on sniffing the user language
and instead always redirect to the default language.